Reinitialized nextjs app
