+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "clsx": "^2.1.0",
     "framer-motion": "^11.0.25",
-    "next": "14.1.4",
+    "next": "^14.2.3",
     "react": "^18",
     "react-dom": "^18",
     "tailwind-merge": "^2.2.2"
